Study Selection Menu Changed
Added buttons for Optimized Thigh Extension Study and Aim 2 Thigh Extension Study trials. Also added the ability to scroll since we now have too many buttons for the length of the screen. Set String resources for the Study Selection Menu to get rid of the warnings in the layout file.